added validation

+++ src/main/java/com/alexandruleonte/api/ChapterRest.java
@@ -2,6 +2,8 @@ package com.alexandruleonte.api;
 
 import com.alexandruleonte.dao.ChapterDao;
 import com.alexandruleonte.entities.Chapter;
+import com.alexandruleonte.entities.Platform;
+import com.alexandruleonte.service.MapValidationErrorService;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
@@ -14,6 +16,9 @@ public class ChapterRest {
     @Inject
     ChapterDao chapterDao;
 
+    @Inject
+    MapValidationErrorService errorService;
+
     @GET
     public Response chapters() {
         return Response.ok(chapterDao.getChapters()).build();
@@ -22,11 +27,18 @@ public class ChapterRest {
     @GET
     @Path("{id}")
     public Response getChapterById(@PathParam("id") int id) {
-        return Response.ok(chapterDao.getChapter(id)).build();
+        Chapter c = chapterDao.getChapter(id);
+        if (c == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Chapter ID '" + id + "' does not exist").build();
+        }
+        return Response.ok(c).build();
     }
 
     @POST
     public Response createChapter(Chapter chapter) {
+        Response errorMap = errorService.getErrorMap(chapter);
+        if (errorMap != null) return errorMap;
+
         chapterDao.save(chapter);
         return Response.status(Response.Status.CREATED).build();
     }
@@ -34,6 +46,10 @@ public class ChapterRest {
     @DELETE
     @Path("{id}")
     public Response deleteChapter(@PathParam("id") int id) {
+        Chapter c = chapterDao.getChapter(id);
+        if (c == null) {
+            return Response.status(Response.Status.NOT_FOUND).entity("Chapter ID '" + id + "' does not exist").build();
+        }
         chapterDao.delete(chapterDao.getChapter(id));
         return Response.ok().build();
     }
